dev/requirements.txt
pydantic==2.8.2
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1